[truetype] Rework bytecode rendering mode.
This replaces a large number of confusing boolean constructs that
describe the target rendering mode in the execution context with
a straight mode variable.  It fixes some hidden bugs when
FT_LOAD_TARGET_XXX were used as flags, which they are not.
The condition that triggers the CV program re-execution is simplified.
These events due the rendering mode change to the are rather rare and
unexpected and, therefore, should not be over-analyzed.

* src/truetype/ttinterp.h (TT_ExecContext): Replace 'grayscale',
'grayscale_cleartype', 'subpixel_hinting_lean', and 'vertical_lcd_lean'
with the rendering 'mode'.
* src/truetype/ttgload.c (tt_loader_init): Replace 'prep' re-execution
trigger and update 'backward_compatibility' condition.
(tt_loader_set_pp): Updated.
* src/truetype/ttinterp.c (Ins_GETINFO): Updated.