[truetype] Simplify twilight zone management (cont'd).
A hard limit of 256 on the number of the twilight zone points is
imposed. Microsoft fonts use 16 typically. This should be enough
for various metric anchors, which si the primary purpose of the
twilight zone.  We also deduplicate the zone and place it only in
the execution context structure.

* src/sfnt/ttload.c (tt_face_load_maxp): Limit 'maxTwilightPoints'.
* src/truetype/ttinterp.c (TT_Load_Context, TT_Run_Context): Do not
tweak or handle the twilight zone.
* src/truetype/ttobjs.c (tt_size_init_bytecodea, tt_size_done_bytecode,
tt_size_run_prep): Updated to set the twilight zone in TT_ExecContext.
* src/truetype/ttobjs.h (TT_Size): Remove 'twilight'.